Cleanup IContentManager
