+++ src/OrchardCore.Modules/OrchardCore.Search.Lucene/Services/LuceneIndexingService.cs
@@ -126,9 +126,9 @@ namespace OrchardCore.Search.Lucene
                     var allPublished = new Dictionary<string, ContentItem>();
                     var allLatest = new Dictionary<string, ContentItem>();
 
-                    var allPublishedContentItems = await contentManager.GetAsync(updatedContentItemIds);
+                    var allPublishedContentItems = await contentManager.GetAsync(updatedContentItemIds, VersionOptions.Published);
                     allPublished = allPublishedContentItems.DistinctBy(x => x.ContentItemId).ToDictionary(k => k.ContentItemId, v => v);
-                    var allLatestContentItems = await contentManager.GetAsync(updatedContentItemIds, latest: true);
+                    var allLatestContentItems = await contentManager.GetAsync(updatedContentItemIds, VersionOptions.Latest);
                     allLatest = allLatestContentItems.DistinctBy(x => x.ContentItemId).ToDictionary(k => k.ContentItemId, v => v);
 
                     // Group all DocumentIndex by index to batch update them.
