+++ src/OrchardCore/OrchardCore.ContentManagement.Abstractions/IContentManager.cs
@@ -1,7 +1,4 @@
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text.Json.Nodes;
 using System.Threading.Tasks;
 using OrchardCore.ContentManagement.Handlers;
 
@@ -33,7 +30,6 @@ namespace OrchardCore.ContentManagement
         /// <param name="contentItem">The content instance filled with all necessary data.</param>
         /// <param name="options">The version to create the item with.</param>
         Task CreateAsync(ContentItem contentItem, VersionOptions options);
-
 
         /// <summary>
         /// Creates (puts) a new content item and manages removing and updating existing published or draft items.
@@ -71,28 +67,11 @@ namespace OrchardCore.ContentManagement
         Task<ContentValidateResult> RestoreAsync(ContentItem contentItem);
 
         /// <summary>
-        /// Gets the published content item with the specified id.
-        /// </summary>
-        /// <param name="id">The content item id to load.</param>
-        Task<ContentItem> GetAsync(string id);
-
-        /// <summary>
         /// Gets the content item with the specified id and version.
         /// </summary>
-        /// <param name="id">The id content item id to load.</param>
+        /// <param name="contentItemId">The id of the content item to load.</param>
         /// <param name="options">The version option.</param>
-        Task<ContentItem> GetAsync(string id, VersionOptions options);
-
-        /// <summary>
-        /// Gets the published content items with the specified ids.
-        /// </summary>
-        /// <param name="contentItemIds">The content item ids to load.</param>
-        /// <param name="latest">Whether a draft should be loaded if available. <c>false</c> by default.</param>
-        /// <remarks>
-        /// This method will always issue a database query.
-        /// This means that it should be used only to get a list of content items that have not been loaded.
-        /// </remarks>
-        Task<IEnumerable<ContentItem>> GetAsync(IEnumerable<string> contentItemIds, bool latest = false);
+        Task<ContentItem> GetAsync(string contentItemId, VersionOptions options);
 
         /// <summary>
         /// Gets the published content items with the specified ids.
@@ -138,7 +117,9 @@ namespace OrchardCore.ContentManagement
         Task SaveDraftAsync(ContentItem contentItem);
 
         Task PublishAsync(ContentItem contentItem);
+
         Task UnpublishAsync(ContentItem contentItem);
+
         Task<TAspect> PopulateAspectAsync<TAspect>(IContent content, TAspect aspect);
 
         /// <summary>
@@ -147,139 +128,5 @@ namespace OrchardCore.ContentManagement
         /// <param name="contentItem">The content item to clone.</param>
         /// <returns>Clone of the item.</returns>
         Task<ContentItem> CloneAsync(ContentItem contentItem);
-    }
-
-    public static class ContentManagerExtensions
-    {
-        /// <summary>
-        /// Creates (persists) a new Published content item.
-        /// </summary>
-        /// <param name="contentManager">The <see cref="IContentManager"/> instance.</param>
-        /// <param name="contentItem">The content instance filled with all necessary data.</param>
-        public static Task CreateAsync(this IContentManager contentManager, ContentItem contentItem)
-        {
-            return contentManager.CreateAsync(contentItem, VersionOptions.Published);
-        }
-
-        public static Task<TAspect> PopulateAspectAsync<TAspect>(this IContentManager contentManager, IContent content) where TAspect : new()
-        {
-            return contentManager.PopulateAspectAsync(content, new TAspect());
-        }
-
-        public static async Task<bool> HasPublishedVersionAsync(this IContentManager contentManager, IContent content)
-        {
-            if (content.ContentItem == null)
-            {
-                return false;
-            }
-
-            return content.ContentItem.IsPublished() || (await contentManager.GetAsync(content.ContentItem.ContentItemId, VersionOptions.Published) != null);
-        }
-
-        public static Task<ContentItemMetadata> GetContentItemMetadataAsync(this IContentManager contentManager, IContent content)
-        {
-            return contentManager.PopulateAspectAsync<ContentItemMetadata>(content);
-        }
-
-        public static async Task<IEnumerable<ContentItem>> LoadAsync(this IContentManager contentManager, IEnumerable<ContentItem> contentItems)
-        {
-            var results = new List<ContentItem>(contentItems.Count());
-
-            foreach (var contentItem in contentItems)
-            {
-                results.Add(await contentManager.LoadAsync(contentItem));
-            }
-
-            return results;
-        }
-
-        public static async IAsyncEnumerable<ContentItem> LoadAsync(this IContentManager contentManager, IAsyncEnumerable<ContentItem> contentItems)
-        {
-            await foreach (var contentItem in contentItems)
-            {
-                yield return await contentManager.LoadAsync(contentItem);
-            }
-        }
-
-        public static async Task<ContentValidateResult> UpdateValidateAndCreateAsync(this IContentManager contentManager, ContentItem contentItem, VersionOptions options)
-        {
-            await contentManager.UpdateAsync(contentItem);
-            var result = await contentManager.ValidateAsync(contentItem);
-
-            if (result.Succeeded)
-            {
-                await contentManager.CreateAsync(contentItem, options);
-            }
-
-            return result;
-        }
-
-        /// <summary>
-        /// Gets either the published container content item with the specified id, or if the json path supplied gets the contained content item. 
-        /// </summary>
-        /// <param name="contentManager">The <see cref="IContentManager"/> instance.</param>
-        /// <param name="id">The content item id to load.</param>
-        /// <param name="jsonPath">The json path of the contained content item.</param>
-        public static Task<ContentItem> GetAsync(this IContentManager contentManager, string id, string jsonPath)
-        {
-            return contentManager.GetAsync(id, jsonPath, VersionOptions.Published);
-        }
-
-        /// <summary>
-        /// Gets either the container content item with the specified id and version, or if the json path supplied gets the contained content item.
-        /// </summary>
-        /// <param name="contentManager">The <see cref="IContentManager"/> instance.</param>
-        /// <param name="id">The id content item id to load.</param>
-        /// <param name="options">The version option.</param>
-        /// <param name="jsonPath">The json path of the contained content item.</param>
-        public static async Task<ContentItem> GetAsync(this IContentManager contentManager, string id, string jsonPath, VersionOptions options)
-        {
-            var contentItem = await contentManager.GetAsync(id, options);
-
-            // It represents a contained content item
-            if (!string.IsNullOrEmpty(jsonPath))
-            {
-                var root = (JsonObject)contentItem.Content;
-                contentItem = root.SelectNode(jsonPath)?.ToObject<ContentItem>();
-
-                return contentItem;
-            }
-
-            return contentItem;
-        }
-    }
-
-    public class VersionOptions
-    {
-        /// <summary>
-        /// Gets the latest version.
-        /// </summary>
-        public static VersionOptions Latest { get { return new VersionOptions { IsLatest = true }; } }
-
-        /// <summary>
-        /// Gets the latest published version.
-        /// </summary>
-        public static VersionOptions Published { get { return new VersionOptions { IsPublished = true }; } }
-
-        /// <summary>
-        /// Gets the latest draft version.
-        /// </summary>
-        public static VersionOptions Draft { get { return new VersionOptions { IsDraft = true }; } }
-
-        /// <summary>
-        /// Gets the latest version and creates a new version draft based on it.
-        /// </summary>
-        public static VersionOptions DraftRequired { get { return new VersionOptions { IsDraft = true, IsDraftRequired = true }; } }
-
-        /// <summary>
-        /// Gets all versions.
-        /// </summary>
-        public static VersionOptions AllVersions { get { return new VersionOptions { IsAllVersions = true }; } }
-
-        public bool IsLatest { get; private set; }
-        public bool IsPublished { get; private set; }
-        public bool IsDraft { get; private set; }
-        public bool IsDraftRequired { get; private set; }
-        public bool IsAllVersions { get; private set; }
     }
 }
