+++ src/OrchardCore/OrchardCore.ContentManagement/DefaultContentManager.cs
@@ -83,86 +83,6 @@ namespace OrchardCore.ContentManagement
             return context3.ContentItem;
         }
 
-        public Task<ContentItem> GetAsync(string contentItemId)
-        {
-            return GetAsync(contentItemId, VersionOptions.Published);
-        }
-
-        public async Task<IEnumerable<ContentItem>> GetAsync(IEnumerable<string> contentItemIds, bool latest = false)
-        {
-            ArgumentNullException.ThrowIfNull(contentItemIds);
-
-            var itemIds = contentItemIds
-                .Where(id => id is not null)
-                .Distinct()
-                .ToArray();
-
-            if (itemIds.Length == 0)
-            {
-                return [];
-            }
-
-            List<ContentItem> contentItems = null;
-            List<ContentItem> storedItems = null;
-            if (latest)
-            {
-                contentItems = (await _session
-                    .Query<ContentItem, ContentItemIndex>()
-                    .Where(i => i.ContentItemId.IsIn(itemIds) && i.Latest == true)
-                    .ListAsync()
-                    ).ToList();
-            }
-            else
-            {
-                foreach (var itemId in itemIds)
-                {
-                    // If the published version is already stored, we can return it.
-                    if (_contentManagerSession.RecallPublishedItemId(itemId, out var contentItem))
-                    {
-                        storedItems ??= [];
-                        storedItems.Add(contentItem);
-                    }
-                }
-
-                // Only query the ids not already stored.
-                var itemIdsToQuery = storedItems is not null
-                    ? itemIds.Except(storedItems.Select(c => c.ContentItemId)).ToArray()
-                    : itemIds;
-
-                if (itemIdsToQuery.Length > 0)
-                {
-                    contentItems = (await _session
-                       .Query<ContentItem, ContentItemIndex>()
-                       .Where(i => i.ContentItemId.IsIn(itemIdsToQuery) && i.Published == true)
-                       .ListAsync()
-                       ).ToList();
-                }
-            }
-
-            if (contentItems is not null)
-            {
-                for (var i = 0; i < contentItems.Count; i++)
-                {
-                    contentItems[i] = await LoadAsync(contentItems[i]);
-                }
-
-                if (storedItems is not null)
-                {
-                    contentItems.AddRange(storedItems);
-                }
-            }
-            else if (storedItems is not null)
-            {
-                contentItems = storedItems;
-            }
-            else
-            {
-                return [];
-            }
-
-            return contentItems.OrderBy(c => Array.IndexOf(itemIds, c.ContentItemId));
-        }
-
         public async Task<ContentItem> GetAsync(string contentItemId, VersionOptions options)
         {
             if (string.IsNullOrEmpty(contentItemId))
