+++ src/docs/releases/2.0.0.md
@@ -194,6 +194,11 @@ public class RegisterUserFormDisplayDriver : DisplayDriver<RegisterUserForm>
     }
 }
 ```
+
+### Contents
+
+The `IContentManager` interface was modified. The method `Task<IEnumerable<ContentItem>> GetAsync(IEnumerable<string> contentItemIds, bool latest = false)` was removed. Instead use the method that accepts `VersionOptions` by providing either `VersionOptions.Latest` or `VersionOptions.Published`.
+
 ### Media Indexing
 
 Previously, `.pdf` files were automatically indexed in the search providers (Elasticsearch, Lucene or Azure AI Search). Now, if you want to continue to index `.PDF` file you'll need to enable the `OrchardCore.Media.Indexing.Pdf` feature.
